requirements.txt


docx2txt==0.8
fireworks-ai==0.15.9
langchain==0.3.8
langchain-chroma==0.1.4
langchain-community==0.3.8
langchain-core==0.3.21
langchain-fireworks==0.2.5
Markdown==3.7
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
pdf2image==1.17.0
pdf2text==1.0.0
pdfminer==20191125
pdfminer.six==20231228
pi_heif==0.20.0
pillow==11.0.0
pypdf==5.1.0
PyPDF2==3.0.1
python-docx==1.1.2
streamlit==1.40.2
unstructured==0.16.8
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13


